pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fipsagents-platform"
version = "0.1.0"
description = "Cross-agent platform service for fips-agents deployments"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "Red Hat AI Americas"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
    "Operating System :: OS Independent",
]
keywords = ["ai", "agent", "fips", "openshift", "fastapi"]
dependencies = [
    "fipsagents[feedback,server]>=0.12.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "httpx",
]

[project.urls]
Homepage = "https://github.com/fips-agents/fipsagents-platform"
Repository = "https://github.com/fips-agents/fipsagents-platform"
Issues = "https://github.com/fips-agents/fipsagents-platform/issues"

[project.scripts]
fipsagents-platform = "fipsagents_platform.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
